refactor(network)
