+++ packages/core/src/lib/network/index.ts
@@ -1,2 +1,1 @@
 export * from './network.service';
-export * from './network.module';
