--- packages/core/src/lib/network/network.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { NetworkService } from './network.service';
-
-@NgModule({
-  declarations: [],
-  imports: [CommonModule],
-  providers: [NetworkService]
-})
-export class NetworkModule { }
